# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Sales AI System API
  version: 1.0.0
  description: Browser clients authenticate with an opaque server-side session cookie. Unsafe authenticated requests also require the X-CSRF-Token returned by GET /api/auth/me.
servers:
  - url: http://localhost:3000
security:
  - cookieSession: []
paths:
...
  /api/mails:
    get:
      summary: List mails
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/EmailStatus' }
      responses:
        '200':
          description: Mail list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/draft:
    post:
      summary: Create a mail draft
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMailDraftRequest' }
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/templates:
    get:
      summary: List active mail templates
      parameters:
        - name: channel
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Template response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    post:
      summary: Save or update a mail template
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveMailTemplateRequest' }
      responses:
        '201':
          description: Template response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/templates/{key}:
    get:
      summary: Get a mail template by key
      parameters:
        - $ref: '#/components/parameters/TemplateKey'
      responses:
        '200':
          description: Template response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/templates/import:
    post:
      summary: Import mail templates
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportMailTemplatesRequest' }
      responses:
        '201':
          description: Template import response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}:
    patch:
      summary: Update mail subject or body
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMailRequest' }
      responses:
        '200':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/consistency:
    get:
      summary: Check draft consistency
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Consistency response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/checklist:
    get:
      summary: Get mail checklist
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Checklist response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    patch:
      summary: Update mail checklist
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMailChecklistRequest' }
      responses:
        '200':
          description: Checklist response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/request-review:
    post:
      summary: Request mail review
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/request-rereview:
    post:
      summary: Request mail re-review
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/approve:
    post:
      summary: Approve a mail
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/reject:
    post:
      summary: Reject a mail
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectMailRequest' }
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/queue:
    post:
      summary: Queue an approved mail
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/mark-sent:
    post:
      summary: Mark a mail as sent after manual recovery
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkMailSentRequest' }
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/send:
    post:
      summary: Send a queued mail through the configured provider
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/replies:
    post:
      summary: Record a reply for a mail
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMailReplyRequest' }
      responses:
        '201':
          description: Reply response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/retry:
    post:
      summary: Retry a mail workflow
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/{id}/cancel:
    post:
      summary: Cancel a mail
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Mail response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/mails/threads/{gmailThreadId}:
    get:
      summary: Get a Gmail thread
      parameters:
        - $ref: '#/components/parameters/GmailThreadId'
      responses:
        '200':
          description: Thread response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: sales_ai_session
      description: Opaque server-managed session. Production uses the __Host-sales_ai_session cookie name.
...
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    GmailThreadId:
      name: gmailThreadId
      in: path
      required: true
      schema: { type: string }
    TemplateKey:
      name: key
      in: path
      required: true
      schema: { type: string }
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Required on every unsafe authenticated request. Obtain it from GET /api/auth/me and keep it in memory only.
      schema: { type: string }
  schemas:
    ApiResponse:
      type: object
      required: [data, meta, error]
      properties:
        data: {}
        meta:
          type: object
          nullable: true
          additionalProperties: true
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ErrorResponse'
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    EmailStatus:
      type: string
      enum: [draft, in_review, rejected, approved, queued, sending, sent, failed, cancelled]
...
    SaveMailTemplateRequest:
      type: object
      required: [key, name, body]
      properties:
        key: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        channel: { type: string }
        subject: { type: string }
        body: { type: string, minLength: 1 }
        description: { type: string }
        isActive: { type: boolean }
    ImportMailTemplatesRequest:
      type: object
      required: [templates]
      properties:
        templates:
          type: array
          items: { $ref: '#/components/schemas/SaveMailTemplateRequest' }
    CreateMailDraftRequest:
      type: object
      required: [leadId, analysisRevisionId, templateKey]
      properties:
        leadId: { type: string, format: uuid }
        analysisRevisionId: { type: string, format: uuid }
        templateKey: { type: string, minLength: 1 }
        manualInstruction: { type: string }
    UpdateMailRequest:
      type: object
      properties:
        subject: { type: string }
        body: { type: string }
    UpdateMailChecklistRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UpdateMailChecklistItem'
    UpdateMailChecklistItem:
      type: object
      required: [key, label, checked]
      properties:
        key: { type: string, minLength: 1 }
        label: { type: string, minLength: 1 }
        checked: { type: boolean }
    RejectMailRequest:
      type: object
      properties:
        reason: { type: string }
    MarkMailSentRequest:
      type: object
      properties:
        sentAt: { type: string, format: date-time }
    CreateMailReplyRequest:
      type: object
      required: [body]
      properties:
        fromEmail: { type: string }
        body: { type: string, minLength: 1 }
        receivedAt: { type: string, format: date-time }
...